:sparkles: feature: s3Config 수정
- partial registration 되도록 수정
